ml/requirements.txt
flask
scikit-learn
flask_cors
openai
Pillow
python-dotenv
cloudinary